backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Supabase (handles database connections for us)
supabase==2.10.0

# Environment variables
python-dotenv==1.0.1

# Pydantic for data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Date/time handling
python-dateutil==2.9.0

# HTTP requests - version compatible with supabase
httpx==0.27.2

# Password hashing
passlib[bcrypt]==1.7.4

# JWT tokens
python-jose[cryptography]==3.3.0